Background Image and Jumbotron adjustment
So I got the background image added.  Fifty percent transparent seems to
be nice for reading text while still conveying mood.  I also removed the
jumbotron div and replaced it with a regular row since it was hidden
after adding the image background code.  I tried different placements of
the code to fix, to no avail.  It's probably a CSS conflict.  Just
adding a fresh DIV at the top recovered the text and image without all
the hassle.